detach always returns void
This commit also adds the concept of "attach" so that we can keep attached state
in sync with the parent. The notion of "children" is still a bit fuzzy but the
notion of "parent" is more clear -- a widget can't be attached unless its parent is.

We can use this attached information to optimize the Conditional widget, among others,
to only do dom manip when attached.

This concept needs to be fleshed out a bit more and separated from other kinds of
"children" -- e.g. should a detach remove from a child array? Should destroying a
parent destroy all children and placeholders? All attached widgets?