Make built-in extension events propagate properly
Also reduces the number of hit detections per event to 1 regardless
of how many times the event has been registered.

There should be a better way to do this in the future without
needing to use a confusingly different `type` on the event object;
currently with the string map from dom/Widget#on to actions, an
attempt to use the same name as the type will cause an infinite
circular reference.